% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_group_fun.R
\name{are_group_fun}
\alias{are_group_fun}
\title{Helper function to generate acute respiratory episodes}
\usage{
are_group_fun(var, thresh_d)
}
\arguments{
\item{var}{variable name}

\item{thresh_d}{Threshold}
}
\value{
...
}
\description{
are_group_fun() is used in lk_symp_inter
}
\examples{
are_group_fun()
}
