% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_krank_klass.R
\name{lk_krank_klass}
\alias{lk_krank_klass}
\title{Classification into acute respiratory A- and B-symptoms of symptom diary entries}
\usage{
lk_krank_klass(
  lk_dat,
  diag_num,
  symp_a_vec,
  symp_resp_weitere_vec,
  symp_b_lib_vec,
  write_table = F
)
}
\arguments{
\item{lk_dat}{Input data (takes output from lk_klass)}

\item{diag_num}{Vector of the Diagnose_Symptome-columns to be considered in the symptom diary}

\item{symp_a_vec}{Vector of A-Symptoms (conservative scenario)}

\item{symp_resp_weitere_vec}{Vector of additional A-Symptoms to be considered in the liberal scenario}

\item{symp_b_lib_vec}{Vector of B-Symptoms to be considered in the liberal scenario}

\item{write_table}{Should the output be written to a xlsx-table (TRUE or FALSE)}
}
\value{
A data.frame including the output from lk_klass as well as the additional binary classification columns symp_a_lib_klass (A-Symptom, liberal scenario), symp_a_kons_klass (A-Symptom, conservative scenario) as well as n_symp_b_lib and n_symp_b_kons (number of B-Symptoms in liberal and conservative scenario, respectively)
}
\description{
lk_krank_klass() classifies each record in the symptom diary into A-symptoms, and counts the number of B-symptoms, according to two scenarios (conservative and liberal).
}
\examples{
lk_krank_klass()
}
\keyword{diary}
\keyword{symptom}
