% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_symp_episod.R
\name{lk_symp_episod}
\alias{lk_symp_episod}
\title{Generating episodes data set}
\usage{
lk_symp_episod(lk_dat, scenario, write_table = T)
}
\arguments{
\item{lk_dat}{Input data (takes output from lk_symp_inter)}

\item{scenario}{liberal or conservative scenario}

\item{write_table}{Should the output be written to a xlsx-table (TRUE or FALSE)}
}
\value{
A data.frame containing start and end date as well as duration in days for each ARE
}
\description{
lk_symp_episod() generates episodes data set comprised of time spans for each ARE/ARE-Type.
}
\examples{
lk_symp_episod()
}
\keyword{diary}
\keyword{symptom}
