% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_symp_inter.R
\name{lk_symp_inter}
\alias{lk_symp_inter}
\title{Generates acute respiratory episodes (ARE)}
\usage{
lk_symp_inter(lk_dat, thresh_inter, thresh_d, scenario, write_table = T)
}
\arguments{
\item{lk_dat}{Input data (takes output from lk_krank_klass)}

\item{thresh_inter}{...}

\item{thresh_d}{threshold for gaps}

\item{scenario}{liberal or conservative scenario}

\item{write_table}{Should the output be written to a xlsx-table (TRUE or FALSE)}
}
\value{
A data.frame including the output from lk_klass as well as the additional columns for ARE and ARE-Type
}
\description{
lk_symp_inter() classifies each entry in the symptom diary to a acute respiratory episodes (ARE).
}
\examples{
lk_symp_inter()
}
\keyword{diary}
\keyword{symptom}
