% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_symp_outcome.R
\name{lk_symp_outcome}
\alias{lk_symp_outcome}
\title{Calculation of outcome variables}
\usage{
lk_symp_outcome(lk_inter_dat, lk_epi_dat, scenario, lebmon, write_table = T)
}
\arguments{
\item{lk_inter_dat}{First input data (takes output from lk_symp_inter)}

\item{lk_epi_dat}{Second input data (takes output from lk_epi_dat)}

\item{scenario}{Liberal or conservative scenario}

\item{lebmon}{Duration of months of life considered in the analysis (e.g. 12 months)}

\item{write_table}{Should the output be written to a xlsx-table (TRUE or FALSE)}
}
\value{
A data.frame containing start and end date as well as duration in days for each ARE
}
\description{
lk_symp_outcome() generates a set of outcome variables.
}
\examples{
lk_symp_outcome()
}
\keyword{diary}
\keyword{symptom}
